Use original refererence ordinal where deriveable
With JavaScript enabled, the original order of the references in the list
is stored and used against each reference, both in the main list, and in
the reference popups. This will be necessary when the sort order is
changed. Without js enabled, or if the list of original ordinals can't be
derived, the main list reference numbers come from the ordered list, and
the inline reference links are regular anchor links. This does not
address the use case of having a reference link on a fragment page where
the references are not available.